requirements.txt
flask-jwt-extended
Flask
pandas
deep-translator
razorpay
reportlab
Flask-Limiter
Flask-CORS
cryptography
python-dotenv
bleach
google-genai
mysql-connector-python